+++ rest/routes/HistoryRoutes.php
@@ -11,8 +11,14 @@ use Firebase\JWT\Key;
  * )
  */
 Flight::route("GET /histories", function(){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::json(Flight::history_service()->get_all());
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
 });
+   
 
 /**
   * @OA\Get(path="/history_by_id", tags={"histories"}, security={{"ApiKeyAuth": {}}},
@@ -26,8 +32,15 @@ Flight::route("GET /histories", function(){
   * )
   */
 Flight::route("GET /history_by_id", function(){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::json(Flight::history_service()->get_by_id(Flight::request()->query['id']));
- });
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
+});
+    
+
 
  /**
   * @OA\Get(path="/histories/{id}", tags={"histories"}, security={{"ApiKeyAuth": {}}},
@@ -36,8 +49,14 @@ Flight::route("GET /history_by_id", function(){
   * )
   */
 Flight::route("GET /histories/@id", function($id){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::json(Flight::history_service()->get_by_id($id));
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
 });
+    
 
  /**
  * @OA\Delete(
@@ -56,9 +75,15 @@ Flight::route("GET /histories/@id", function($id){
  * )
  */
 Flight::route("DELETE /histories/@id", function($id){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::history_service()->delete($id);
     Flight::json(['message'=>"history deleted successfully"]);
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
 });
+   
 
 /**
 * @OA\Post(
@@ -85,12 +110,18 @@ Flight::route("DELETE /histories/@id", function($id){
 * )
 */
 Flight::route("POST /histories", function(){
+  $user = Flight::get('user');
+  if(isset($user)){
     $request= Flight::request()->data->getData();
     Flight::json(['message'=>"history added successfully",
                   'data'=>Flight::history_service()->add($request)
                 ]);
-    
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
 });
+    
+  
 
 /**
  * @OA\Put(
@@ -119,13 +150,19 @@ Flight::route("POST /histories", function(){
  */
 //update
 Flight::route("PUT /histories/@id", function($id){
+  $user = Flight::get('user');
+  if(isset($user)){
     $history = Flight::request()->data->getData();
     #$response=$users_dao->update($user,$id);
     Flight::json(['message'=>"history edit successfully",
                   'data'=>Flight::history_service()->update($history,$id)
                 ]);
-    
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
 });
+    
+
