+++ rest/routes/QuestionRoutes.php
@@ -11,8 +11,14 @@ use Firebase\JWT\Key;
  */
 //works
 Flight::route("GET /questions", function(){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::json(Flight::question_service()->get_all());
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
 });
+    
 
 /**
   * @OA\Get(path="/question_by_id", tags={"questions"}, security={{"ApiKeyAuth": {}}},
@@ -27,8 +33,14 @@ Flight::route("GET /questions", function(){
   */
 //works
 Flight::route("GET /question_by_id", function(){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::json(Flight::question_service()->get_by_id(Flight::request()->query['id']));
- });
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
+});
+    
 
  /**
   * @OA\Get(path="/questions/{id}", tags={"questions"}, security={{"ApiKeyAuth": {}}},
@@ -38,8 +50,14 @@ Flight::route("GET /question_by_id", function(){
   */
 //works
 Flight::route("GET /questions/@id", function($id){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::json(Flight::question_service()->get_by_id($id));
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
 });
+    
 
  /**
  * @OA\Delete(
@@ -58,9 +76,15 @@ Flight::route("GET /questions/@id", function($id){
  * )
  */
 Flight::route("DELETE /questions/@id", function($id){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::question_service()->delete($id);
     Flight::json(['message'=>"Question deleted successfully"]);
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
 });
+   
 
 /**
 * @OA\Post(
@@ -85,12 +109,17 @@ Flight::route("DELETE /questions/@id", function($id){
 * )
 */
 Flight::route("POST /questions", function(){
+  $user = Flight::get('user');
+  if(isset($user)){
     $request= Flight::request()->data->getData();
     Flight::json(['message'=>"Question added successfully",
                   'data'=>Flight::question_service()->add($request)
                 ]);
-    
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
 });
+    
 
 /**
  * @OA\Put(
@@ -117,13 +146,18 @@ Flight::route("POST /questions", function(){
  */
 //update
 Flight::route("PUT /questions/@id", function($id){
+  $user = Flight::get('user');
+  if(isset($user)){
     $question = Flight::request()->data->getData();
     #$response=$users_dao->update($user,$id);
     Flight::json(['message'=>"Question edit successfully",
                   'data'=>Flight::question_service()->update($question,$id)
                 ]);
-    
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
 });
+
