+++ rest/routes/UserRoutes.php
@@ -3,6 +3,7 @@
 use Firebase\JWT\JWT;
 use Firebase\JWT\Key;
 
+
 /**
  * @OA\Get(path="/users", tags={"users"}, security={{"ApiKeyAuth": {}}},
  *         summary="Return all users from the API. ",
@@ -12,12 +13,14 @@ use Firebase\JWT\Key;
 
 //works
 Flight::route('GET /users', function () {
-    
-    //$results=Flight::user_service()->get_all();
-    //print_r($results);
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::json(Flight::user_service()->get_all());
-
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
 });
+
 
 /**
   * @OA\Get(path="/users/{id}", tags={"users"}, security={{"ApiKeyAuth": {}}},
@@ -27,8 +30,15 @@ Flight::route('GET /users', function () {
   */
 //works
 Flight::route('GET /users/@id', function($id){
+  $user = Flight::get('user');
+  
+  if(isset($user)){
     Flight::json(Flight::user_service()->get_by_id($id));
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
 });
+
 
 /**
   * @OA\Get(path="/user_by_id", tags={"users"}, security={{"ApiKeyAuth": {}}},
@@ -44,8 +54,13 @@ Flight::route('GET /users/@id', function($id){
 //http://localhost/Dedsec/rest/user_by_id?id=4
 //works
 Flight::route('GET /user_by_id', function(){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::json(Flight::user_service()->get_by_id(Flight::request()->query['id']));
- });
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
+});
 
   /**
  * @OA\Delete(
@@ -65,8 +80,13 @@ Flight::route('GET /user_by_id', function(){
  */
 //treba staviti status:active or not
 Flight::route('DELETE /users/@id', function($id){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::user_service()->delete($id);
     Flight::json(['message'=>"User deleted successfully"]);
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
 });
 
  /**
@@ -93,14 +113,19 @@ Flight::route('DELETE /users/@id', function($id){
 *     )
 * )
 */
-//works
+//isset($user) && $user['authorization'] == 'authorized' da li je authorized da li je unauthorided ako svi pristupaju ne trebam
 Flight::route('POST /users', function(){
+  $user = Flight::get('user');
+  if(isset($user)){ //dodati ovdje
     $request= Flight::request()->data->getData();
     Flight::json(['message'=>"User added successfully",
                   'data'=>Flight::user_service()->add($request)
                 ]);
-    
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
 });
+   
 
  /**
  * @OA\Put(
@@ -130,12 +155,16 @@ Flight::route('POST /users', function(){
 //update
 //works
 Flight::route('PUT /users/@id', function($id){
+  $user = Flight::get('user');
+  if(isset($user)){
     $user= Flight::request()->data->getData();
     #$response=$users_dao->update($user,$id);
     Flight::json(['message'=>"User edit successfully",
                   'data'=>Flight::user_service()->update($user,$id)
                 ]);
-    
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
 });
 
 /**
@@ -170,7 +199,10 @@ Flight::route('POST /login', function(){
     if (isset($user['userid'])){
       if($user['password'] == md5($login['password'])){
         unset($user['password']);
-        $user['is_admin'] = true;
+        unset($user['surname']);
+        unset($user['email']);
+        unset($user['status']);
+        //$user['is_admin'] = true;
         $jwt = JWT::encode($user, Config::JWT_SECRET(), 'HS256');
         Flight::json(['token' => $jwt]);
       }else{
@@ -211,6 +243,7 @@ Flight::route('POST /users/login', function(){
     }
 });
 */
+
 /**
   * @OA\Get(path="/users/{id}/history", tags={"users"}, security={{"ApiKeyAuth": {}}},
   *     @OA\Parameter(in="path", name="id", example=1, description="User History"),
@@ -218,7 +251,12 @@ Flight::route('POST /users/login', function(){
   * )
   */
 Flight::route('GET /users/@id/history', function($id){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::json(Flight::history_service()->get_by_user_id($id));
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  }
 });
